ProdutoRepository e TesteProdutoRepository implementado

+++ DonaLaura/DBDonaLaura/dbo/TBProduto.sql
@@ -2,8 +2,8 @@
 (
 	[Id] INT IDENTITY (1, 1) NOT NULL, 
     [Nome] VARCHAR(50) NULL, 
-    [PrecoVenda] NUMERIC(10, 2) NULL, 
-    [PrecoCusto] NUMERIC(10, 2) NULL, 
+    [PrecoVenda] FLOAT NULL, 
+    [PrecoCusto] FLOAT NULL, 
     [Disponibilidade] BIT NULL, 
     [DataFabricacao] DATE NULL, 
     [DataValidade] DATE NULL,
